fixed extraction, added logging

+++ sdm/utils/filesmerger.py
@@ -52,6 +52,9 @@ def merge(filename_pattern, output_filename, mode=Mode.txt, batch=20):
 
         files = next_iterable
 
+    if not tempfiles:
+        raise ValueError ("no files to merge")
+
     for tmpfile in tempfiles[:-1]:
         os.remove(tmpfile)
         print("REMOVING", tmpfile)
